Move authentication middleware to backend folder
